move components to separate folder & add navigation components
